Updated Employer Entity and Repository

+++ backend_pat/src/main/java/com/pat/backend_pat/repository/EmployerRepository.java
@@ -1,9 +1,7 @@
 package com.pat.backend_pat.repository;
 
-
-
 import com.pat.backend_pat.entity.Employer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployerRepository extends JpaRepository<Employer, Long> {
+public interface EmployerRepository extends JpaRepository<Employer, Integer> {
 }
